[refactor] components 폴더 더 자세히 목적 별로 분류 + FileStructure를 위한 로직 및 타입, 파일 삭제
