+++ src/components/test/NameInput.test.tsx
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom";
 import { FieldError } from "react-hook-form";
-import NameInput from "../NameInput";
+import NameInput from "../form/NameInput";
 import { render, screen } from "@testing-library/react";
 import React from "react";
