+++ src/components/test/PasswordInput.test.tsx
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom";
-import PasswordInput, { RegisterType } from "../PasswordInput";
+import PasswordInput, { RegisterType } from "../form/PasswordInput";
 import { FieldError } from "react-hook-form";
 import { render, screen } from "@testing-library/react";
 import React from "react";
