+++ src/pages/Main.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SubmitButton from "../components/SubmitButton";
+import SubmitButton from "../components/form/SubmitButton";
 
 function Main() {
   return (
